use version detection support from `_compat`; get rid of `util`
